app/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
pymysql==1.0.3
marshmallow==3.19.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
